Some big changes without breaking the original structure: safer asynchronous execution, done only within the current Tick; more useful methods for the CraftTweaker API; a simple performance analyzer to provide developers with the ability to check performance; fixing a server bug in TheOneProbe; energy units will be formatted (100000 -> 1M); the controller's progress display now supports RecipeModifier modified time; fixes an issue with RecipeModifier where values exceeding the int max when applying RequirementTypeEnergy resulted in missing values.

+++ src/main/java/hellfirepvp/modularmachinery/common/crafting/requirement/RequirementEnergy.java
@@ -58,7 +58,7 @@ public class RequirementEnergy extends ComponentRequirement.PerTick<Long, Requir
 
     @Override
     public ComponentRequirement<Long, RequirementTypeEnergy> deepCopyModified(List<RecipeModifier> modifiers) {
-        int requirement = Math.round(RecipeModifier.applyModifiers(modifiers, this, this.requirementPerTick, false));
+        long requirement = Math.round((double) RecipeModifier.applyModifiers(modifiers, this, this.requirementPerTick, false));
         RequirementEnergy energy = new RequirementEnergy(this.actionType, requirement);
         energy.activeIO = this.activeIO;
         return energy;
